backend/requirements.txt
Jinja2==3.1.6
MarkupSafe==3.0.2
Werkzeug==3.1.3
blinker==1.9.0
click==8.1.8
colorama==0.4.6
flask==3.1.0
itsdangerous==2.2.0
flask_cors==5.0.1
dnspython==2.7.0
pymongo==4.11.3
alpaca-py==0.42.0
yfinance==0.2.62
apscheduler==3.11.0
stable-baselines3==2.7.0
boto3==1.42.71
ta==0.11.0
PyJWT[crypto]==2.10.1
